Simplify the websocket implementation to use a single callback
Instead of using a delegate-style system, this patch moves to use
a callback-event-like system.  This simplifies the code for the
consumer and Implementation.  This will allow multi-threaded
systems much easier by giving a single point where a consumer could
plum the calls out of the websocket API up to another thread.